extracted project date range
